+++ src/main/java/org/sonar/plugins/cas/util/SonarCasProperties.java
@@ -28,7 +28,7 @@ import org.sonar.api.config.Configuration;
  */
 public enum SonarCasProperties {
     /**
-     * Should SONAR create new users? Must be set to TRUE.
+     * Should SonarQube create new users? Must be set to TRUE.
      */
     SONAR_CREATE_USERS("sonar.authenticator.createUsers", SonarPropertyType.BOOLEAN),
     /**
@@ -49,7 +49,7 @@ public enum SonarCasProperties {
     CAS_SERVER_URL_PREFIX("sonar.cas.casServerUrlPrefix", SonarPropertyType.STRING),
     /**
      * Specifies how SonarQube groups should be replicated.
-
+     *
      * <p>A value of <code>CAS</code> always overwrites the user's
      * local groups with the group provided by CAS upon UI log-in. The user's local groups will be untouched with other
      * values.</p>
@@ -92,6 +92,13 @@ public enum SonarCasProperties {
      * Ignore certification validation errors. CAUTION! NEVER USE IN PROD! SECURITY RISK!
      */
     DISABLE_CERT_VALIDATION("sonar.cas.disableCertValidation", SonarPropertyType.BOOLEAN),
+    /**
+     * This value determines whether a Redirect Cookie after the CAS login uses the Secure-flag. Cookies with
+     * <code>secure=true</code> will only work properly in conjunction with HTTPS.
+     *
+     * <p>default to <code>true</code></p>
+     */
+    USE_SECURE_REDIRECT_COOKIES("sonar.cas.userSecureRedirectCookies", SonarPropertyType.BOOLEAN),
     /**
      * The expiration time of the cookie which helps to restore the originally requested SonarQube URL over the CAS authentication
      */